test: dedupe source-tree walker into helpers, tighten fake electron
Hoist collectSourceFiles/SRC_DIR from config.spec.js into test/helpers.js
and reuse in both guard specs; inline the defaultSession getter, drop
unused fake-app scaffolding, and merge the two overlapping regression
assertions into one test.
